Update JUnit test case
