+++ src/test/java/seedu/careflow/logic/CareFlowLogicManagerTest.java
@@ -4,15 +4,15 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static seedu.careflow.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.careflow.commons.core.Messages.MESSAGE_INVALID_DRUG_DISPLAYED_INDEX;
 import static seedu.careflow.commons.core.Messages.MESSAGE_INVALID_PATIENT_DISPLAYED_INDEX;
-import static seedu.careflow.logic.commands.patientcommands.CommandTestUtil.ADDRESS_DESC_AMY;
-import static seedu.careflow.logic.commands.patientcommands.CommandTestUtil.DOB_DESC_AMY;
-import static seedu.careflow.logic.commands.patientcommands.CommandTestUtil.DRUG_ALLERGY_DESC_AMY;
-import static seedu.careflow.logic.commands.patientcommands.CommandTestUtil.EMAIL_DESC_AMY;
-import static seedu.careflow.logic.commands.patientcommands.CommandTestUtil.EMERGENCY_CONTACT_DESC_AMY;
-import static seedu.careflow.logic.commands.patientcommands.CommandTestUtil.GENDER_DESC_AMY;
-import static seedu.careflow.logic.commands.patientcommands.CommandTestUtil.IC_DESC_AMY;
-import static seedu.careflow.logic.commands.patientcommands.CommandTestUtil.NAME_DESC_AMY;
-import static seedu.careflow.logic.commands.patientcommands.CommandTestUtil.PHONE_DESC_AMY;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.ADDRESS_DESC_AMY;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.DOB_DESC_AMY;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.DRUG_ALLERGY_DESC_AMY;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.EMAIL_DESC_AMY;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.EMERGENCY_CONTACT_DESC_AMY;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.GENDER_DESC_AMY;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.IC_DESC_AMY;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.NAME_DESC_AMY;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.PHONE_DESC_AMY;
 import static seedu.careflow.testutil.Assert.assertThrows;
 import static seedu.careflow.testutil.TypicalPatients.AMY;
