+++ src/test/java/seedu/careflow/logic/commands/drugcommands/AddCommandIntegrationTest.java
@@ -1,7 +1,6 @@
 package seedu.careflow.logic.commands.drugcommands;
 
-import static seedu.careflow.logic.commands.CommandTestUtil.assertCommandSuccess;
-import static seedu.careflow.logic.commands.CommandTestUtil.assertDrugCommandFailure;
+import static seedu.careflow.logic.commands.drugcommands.DrugCommandTestUtil.assertDrugCommandFailure;
 import static seedu.careflow.testutil.TypicalDrugs.getTypicalDrugInventory;
 
 import org.junit.jupiter.api.BeforeEach;
@@ -33,7 +32,7 @@ public class AddCommandIntegrationTest {
                 model.getDrugInventory(), new UserPrefs());
         expectedModel.addDrug(validDrug);
 
-        assertCommandSuccess(new AddCommand(validDrug), model,
+        DrugCommandTestUtil.assertCommandSuccess(new AddCommand(validDrug), model,
                 String.format(AddCommand.MESSAGE_SUCCESS, validDrug), expectedModel);
     }
