+++ src/test/java/seedu/careflow/logic/commands/patientcommands/AddCommandTest.java
@@ -4,7 +4,7 @@ import static java.util.Objects.requireNonNull;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static seedu.careflow.logic.commands.patientcommands.CommandTestUtil.VALID_IC_AMY;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.VALID_IC_AMY;
 import static seedu.careflow.testutil.Assert.assertThrows;
 
 import java.nio.file.Path;
