+++ src/test/java/seedu/careflow/logic/commands/patientcommands/DeleteCommandTest.java
@@ -2,9 +2,9 @@ package seedu.careflow.logic.commands.patientcommands;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static seedu.careflow.logic.commands.patientcommands.CommandTestUtil.assertCommandFailure;
-import static seedu.careflow.logic.commands.patientcommands.CommandTestUtil.assertCommandSuccess;
-import static seedu.careflow.logic.commands.patientcommands.CommandTestUtil.showPatientAtIndex;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.assertCommandFailure;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.assertCommandSuccess;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.showPatientAtIndex;
 import static seedu.careflow.testutil.TypicalDrugs.getTypicalDrugInventory;
 import static seedu.careflow.testutil.TypicalIndexes.INDEX_FIRST;
 import static seedu.careflow.testutil.TypicalIndexes.INDEX_SECOND;
