+++ src/test/java/seedu/careflow/logic/commands/patientcommands/ListCommandTest.java
@@ -1,7 +1,7 @@
 package seedu.careflow.logic.commands.patientcommands;
 
-import static seedu.careflow.logic.commands.patientcommands.CommandTestUtil.assertCommandSuccess;
-import static seedu.careflow.logic.commands.patientcommands.CommandTestUtil.showPatientAtIndex;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.assertCommandSuccess;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.showPatientAtIndex;
 import static seedu.careflow.testutil.TypicalDrugs.getTypicalDrugInventory;
 import static seedu.careflow.testutil.TypicalIndexes.INDEX_FIRST;
 import static seedu.careflow.testutil.TypicalPatients.getTypicalPatientRecord;
