+++ src/test/java/seedu/careflow/logic/commands/patientcommands/UpdateCommandTest.java
@@ -2,15 +2,15 @@ package seedu.careflow.logic.commands.patientcommands;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static seedu.careflow.logic.commands.CommandTestUtil.DESC_AMY;
-import static seedu.careflow.logic.commands.CommandTestUtil.DESC_BOB;
-import static seedu.careflow.logic.commands.CommandTestUtil.VALID_ADDRESS_BOB;
-import static seedu.careflow.logic.commands.CommandTestUtil.VALID_DRUG_ALLERGY_BOB;
-import static seedu.careflow.logic.commands.CommandTestUtil.VALID_NAME_BOB;
-import static seedu.careflow.logic.commands.CommandTestUtil.VALID_PHONE_BOB;
-import static seedu.careflow.logic.commands.patientcommands.CommandTestUtil.assertCommandFailure;
-import static seedu.careflow.logic.commands.patientcommands.CommandTestUtil.assertCommandSuccess;
-import static seedu.careflow.logic.commands.patientcommands.CommandTestUtil.showPatientAtIndex;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.DESC_AMY;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.DESC_BOB;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.VALID_ADDRESS_BOB;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.VALID_DRUG_ALLERGY_BOB;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.VALID_NAME_BOB;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.VALID_PHONE_BOB;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.assertCommandFailure;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.assertCommandSuccess;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.showPatientAtIndex;
 import static seedu.careflow.logic.commands.patientcommands.UpdateCommand.MESSAGE_DUPLICATE_PATIENT_NAME;
 import static seedu.careflow.testutil.TypicalDrugs.getTypicalDrugInventory;
 import static seedu.careflow.testutil.TypicalIndexes.INDEX_FIRST;
