+++ src/test/java/seedu/careflow/logic/commands/patientcommands/UpdatePatientDescriptorTest.java
@@ -3,17 +3,17 @@ package seedu.careflow.logic.commands.patientcommands;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static seedu.careflow.logic.commands.patientcommands.CommandTestUtil.DESC_AMY;
-import static seedu.careflow.logic.commands.patientcommands.CommandTestUtil.DESC_BOB;
-import static seedu.careflow.logic.commands.patientcommands.CommandTestUtil.VALID_ADDRESS_BOB;
-import static seedu.careflow.logic.commands.patientcommands.CommandTestUtil.VALID_BIRTHDATE_BOB;
-import static seedu.careflow.logic.commands.patientcommands.CommandTestUtil.VALID_DRUG_ALLERGY_BOB;
-import static seedu.careflow.logic.commands.patientcommands.CommandTestUtil.VALID_EMAIL_BOB;
-import static seedu.careflow.logic.commands.patientcommands.CommandTestUtil.VALID_EMERGENCY_CONTACT_BOB;
-import static seedu.careflow.logic.commands.patientcommands.CommandTestUtil.VALID_GENDER_BOB;
-import static seedu.careflow.logic.commands.patientcommands.CommandTestUtil.VALID_IC_BOB;
-import static seedu.careflow.logic.commands.patientcommands.CommandTestUtil.VALID_NAME_BOB;
-import static seedu.careflow.logic.commands.patientcommands.CommandTestUtil.VALID_PHONE_BOB;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.DESC_AMY;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.DESC_BOB;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.VALID_ADDRESS_BOB;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.VALID_BIRTHDATE_BOB;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.VALID_DRUG_ALLERGY_BOB;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.VALID_EMAIL_BOB;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.VALID_EMERGENCY_CONTACT_BOB;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.VALID_GENDER_BOB;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.VALID_IC_BOB;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.VALID_NAME_BOB;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.VALID_PHONE_BOB;
 
 import org.junit.jupiter.api.Test;
