+++ src/test/java/seedu/careflow/testutil/TypicalPatients.java
@@ -1,23 +1,23 @@
 package seedu.careflow.testutil;
 
-import static seedu.careflow.logic.commands.CommandTestUtil.VALID_ADDRESS_AMY;
-import static seedu.careflow.logic.commands.CommandTestUtil.VALID_ADDRESS_BOB;
-import static seedu.careflow.logic.commands.CommandTestUtil.VALID_DOB_AMY;
-import static seedu.careflow.logic.commands.CommandTestUtil.VALID_DOB_BOB;
-import static seedu.careflow.logic.commands.CommandTestUtil.VALID_DRUG_ALLERGY_AMY;
-import static seedu.careflow.logic.commands.CommandTestUtil.VALID_DRUG_ALLERGY_BOB;
-import static seedu.careflow.logic.commands.CommandTestUtil.VALID_EMAIL_AMY;
-import static seedu.careflow.logic.commands.CommandTestUtil.VALID_EMAIL_BOB;
-import static seedu.careflow.logic.commands.CommandTestUtil.VALID_EMERGENCY_CONTACT_AMY;
-import static seedu.careflow.logic.commands.CommandTestUtil.VALID_EMERGENCY_CONTACT_BOB;
-import static seedu.careflow.logic.commands.CommandTestUtil.VALID_GENDER_AMY;
-import static seedu.careflow.logic.commands.CommandTestUtil.VALID_GENDER_BOB;
-import static seedu.careflow.logic.commands.CommandTestUtil.VALID_IC_AMY;
-import static seedu.careflow.logic.commands.CommandTestUtil.VALID_IC_BOB;
-import static seedu.careflow.logic.commands.CommandTestUtil.VALID_NAME_AMY;
-import static seedu.careflow.logic.commands.CommandTestUtil.VALID_NAME_BOB;
-import static seedu.careflow.logic.commands.CommandTestUtil.VALID_PHONE_AMY;
-import static seedu.careflow.logic.commands.CommandTestUtil.VALID_PHONE_BOB;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.VALID_ADDRESS_AMY;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.VALID_ADDRESS_BOB;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.VALID_BIRTHDATE_AMY;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.VALID_BIRTHDATE_BOB;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.VALID_DRUG_ALLERGY_AMY;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.VALID_DRUG_ALLERGY_BOB;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.VALID_EMAIL_AMY;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.VALID_EMAIL_BOB;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.VALID_EMERGENCY_CONTACT_AMY;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.VALID_EMERGENCY_CONTACT_BOB;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.VALID_GENDER_AMY;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.VALID_GENDER_BOB;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.VALID_IC_AMY;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.VALID_IC_BOB;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.VALID_NAME_AMY;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.VALID_NAME_BOB;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.VALID_PHONE_AMY;
+import static seedu.careflow.logic.commands.patientcommands.PatientCommandTestUtil.VALID_PHONE_BOB;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -65,11 +65,11 @@ public class TypicalPatients {
 
     // Manually added - Person's details found in {@code CommandTestUtil}
     public static final Patient AMY = new PatientBuilder().withName(VALID_NAME_AMY).withPhone(VALID_PHONE_AMY)
-            .withEmail(VALID_EMAIL_AMY).withAddress(VALID_ADDRESS_AMY).withDob(VALID_DOB_AMY)
+            .withEmail(VALID_EMAIL_AMY).withAddress(VALID_ADDRESS_AMY).withDob(VALID_BIRTHDATE_AMY)
             .withGender(VALID_GENDER_AMY).withIc(VALID_IC_AMY).withDrugAllergy(VALID_DRUG_ALLERGY_AMY)
             .withEmergencyContact(VALID_EMERGENCY_CONTACT_AMY).build();
     public static final Patient BOB = new PatientBuilder().withName(VALID_NAME_BOB).withPhone(VALID_PHONE_BOB)
-            .withEmail(VALID_EMAIL_BOB).withAddress(VALID_ADDRESS_BOB).withDob(VALID_DOB_BOB)
+            .withEmail(VALID_EMAIL_BOB).withAddress(VALID_ADDRESS_BOB).withDob(VALID_BIRTHDATE_BOB)
             .withGender(VALID_GENDER_BOB).withIc(VALID_IC_BOB).withDrugAllergy(VALID_DRUG_ALLERGY_BOB)
             .withEmergencyContact(VALID_EMERGENCY_CONTACT_BOB).build();
